+++ src/plugin/gui/panels/gui_library_dev_panel.hpp
@@ -13,7 +13,7 @@ struct LibraryDevPanelContext {
 };
 
 struct LibraryDevPanelState {
-    enum class Tab : u32 {
+    enum class Tab : u8 {
         TagBuilder,
         Utilities,
         Count,
